Line art on the tube: chunked push and a vectoriser
A pushed list was capped at one frame — about 34 items — which is fine for a
banner or a face template and nowhere near enough for a drawing. PushBegin /
PushChunk / PushCommit stage the identical byte stream across as many frames as
it takes, and Commit decodes it with the same parser and the same bounds
checks, so there is no second format to get wrong.

Chunks carry no sequence number on purpose: the link is an ordered, CRC-checked
byte stream, so a chunk cannot arrive out of order, only fail to arrive — and
Commit catches that because the staged bytes then fail to decode.

tools/vec2scene.py turns line art into a scene, from SVG (paths, polylines,
rects, beziers flattened) or from a raster. For raster it traces the CENTRE-
LINE of each stroke rather than the outline of each blob, which matters twice
on pixel art: outlining draws every one-pixel line as a long thin rectangle,
and a staircased diagonal's centre-line simplifies to one clean diagonal, which
is what the beam wants anyway. Douglas-Peucker is tightened automatically until
the drawing fits the device's 128 items, and reports what it did.

Two things this shook out:

- Diagonal adjacency has to require BOTH orthogonal neighbours empty, not
  merely not-both-filled. The looser rule invents diagonals across dense areas
  and the cat's whiskers came out as a cat's cradle of little triangles.

- Frames near MAX_PAYLOAD were being dropped silently. HWCDC's TX ring
  defaults to 256 bytes, so a 244-byte frame all but fills it in one call and
  the tail is discarded rather than queued. Every frame that had ever worked
  was tiny — ping, set-time, banner — so this only appeared once there was
  artwork to send. The ring is now 2KB and chunks are 96 bytes, comfortably
  clear of the edge.

Verified end to end: 28 segments staged as 96+96+61 bytes, committed, decoded,
and the device reports mode=pushed.

+++ display-teensy/include/drawlist.h
@@ -59,5 +59,5 @@ void expandTemplate(DrawList& list, const ClockState& clk,
 // everything: item count, per-item length against the remaining payload, and
 // arena space. On any malformed input it returns false having emptied `out`,
 // which blanks the display rather than rendering something half-parsed.
-bool decodePushList(const uint8_t* payload, uint8_t len,
+bool decodePushList(const uint8_t* payload, uint16_t len,
                     DrawList& out, char* arena, uint16_t arenaCap);
